Extract option types into file
