+++ packages/model-serving/src/components/deploymentWizard/fields/ModelLocationInputFields.tsx
@@ -30,7 +30,7 @@ import NewConnectionField from './modelLocationFields/NewConnectionField';
 import { PvcSelectField } from './modelLocationFields/PVCSelectField';
 import { CustomTypeSelectField } from './modelLocationFields/CustomTypeSelectField';
 import { useEnabledModelServingConnectionTypes } from './modelLocationFields/useEnabledConnectionTypes';
-import { ociOption, s3Option, uriOption } from './ModelLocationSelectField';
+import { ociOption, s3Option, uriOption } from './modelLocationFields/modeLocationTypes';
 import usePvcs from '../../../concepts/usePvcs';
 import { ModelLocationData, ModelLocationType } from '../types';
 import { resolveConnectionType } from '../utils';
