+++ packages/model-serving/src/components/deploymentWizard/fields/modelLocationFields/modeLocationTypes.ts
@@ -1,0 +1,12 @@
+export const s3Option = {
+  key: 'S3',
+  label: 'S3 object storage',
+};
+export const ociOption = {
+  key: 'OCI',
+  label: 'OCI compliant registry',
+};
+export const uriOption = {
+  key: 'URI',
+  label: 'URI',
+};
